1. Added Finder functionality (from previous app version). 2. Enabled processing changes for skirmish missions. 3. Disabled possibility to remove or change "vanilla" rebalance config. 4. Enabled tab for changing "Other" addresses. 5. Changed app background image.

+++ SHC-Rebalancer/Modules/Main/MainContext.cs
@@ -1,4 +1,5 @@
-﻿using System.IO;
+﻿using System.Collections.ObjectModel;
+using System.IO;
 using System.Reflection;
 using System.Windows.Controls;
 
@@ -11,6 +12,7 @@ public class MainContext : StswObservableObject
     public StswCommand<object> ReloadRebalancesCommand { get; }
     public StswAsyncCommand<string> EditRebalanceCommand { get; }
     public StswAsyncCommand<string> RemoveRebalanceCommand { get; }
+    public StswAsyncCommand<GameVersion> FindCommand { get; }
 
     public MainContext()
     {
@@ -20,7 +22,9 @@ public class MainContext : StswObservableObject
 
         ReloadRebalancesCommand = new(ReloadRebalances);
         EditRebalanceCommand = new(EditRebalance);
-        RemoveRebalanceCommand = new(RemoveRebalance);
+        RemoveRebalanceCommand = new(RemoveRebalance, () => Settings.Default.RebalanceName != "vanilla");
+
+        FindCommand = new(Find);
 
         Task.Run(() => ReloadRebalancesCommand.Execute());
     }
@@ -148,9 +152,28 @@ public class MainContext : StswObservableObject
             if (string.IsNullOrWhiteSpace(parameter))
                 return;
 
+            if (parameter == "vanilla")
+            {
+                await StswMessageDialog.Show("`vanilla` rebalance config cannot be removed.", "Information", null, StswDialogButtons.OK, StswDialogImage.Information);
+                return;
+            }
+
             var filePath = Path.Combine(Storage.PathRebalances, parameter + ".json");
             if (File.Exists(filePath) && await StswMessageDialog.Show($"Are you sure you want to remove '{parameter}' rebalance config?", "Confirmation", null, StswDialogButtons.YesNo, StswDialogImage.Question) == true)
                 File.Delete(filePath);
+        }
+        catch (Exception ex)
+        {
+            await StswMessageDialog.Show(ex, MethodBase.GetCurrentMethod()?.Name, true);
+        }
+    }
+
+    /// Find
+    public async Task Find(GameVersion type)
+    {
+        try
+        {
+            Finder.Find(FinderResults, type, FinderFilterSize, FinderFilterAddress, FinderFilterSkips, FinderFilterValues);
         }
         catch (Exception ex)
         {
@@ -175,4 +198,64 @@ public class MainContext : StswObservableObject
         set => SetProperty(ref _selectedRebalance, value);
     }
     private RebalanceModel? _selectedRebalance;
+
+    /// FinderFilterType
+    public GameVersion? FinderFilterType
+    {
+        get => _finderFilterType;
+        set => SetProperty(ref _finderFilterType, value);
+    }
+    private GameVersion? _finderFilterType;
+
+    /// FinderFilterSize
+    public int FinderFilterSize
+    {
+        get => _finderFilterSize;
+        set => SetProperty(ref _finderFilterSize, value);
+    }
+    private int _finderFilterSize = 1;
+
+    /// FinderFilterAddress
+    public string? FinderFilterAddress
+    {
+        get => _finderFilterAddress;
+        set
+        {
+            SetProperty(ref _finderFilterAddress, value);
+            FindCommand.Execute(FinderFilterType);
+        }
+    }
+    private string? _finderFilterAddress;
+
+    /// FinderFilterSkips
+    public int FinderFilterSkips
+    {
+        get => _finderFilterSkips;
+        set
+        {
+            SetProperty(ref _finderFilterSkips, value);
+            FindCommand.Execute(FinderFilterType);
+        }
+    }
+    private int _finderFilterSkips = 0;
+
+    /// FinderFilterValues
+    public string? FinderFilterValues
+    {
+        get => _finderFilterValues;
+        set
+        {
+            SetProperty(ref _finderFilterValues, value);
+            FindCommand.Execute(FinderFilterType);
+        }
+    }
+    private string? _finderFilterValues;
+
+    /// FinderResults
+    public ObservableCollection<FinderDataModel> FinderResults
+    {
+        get => _finderResults;
+        set => SetProperty(ref _finderResults, value);
+    }
+    private ObservableCollection<FinderDataModel> _finderResults = [];
 }
